% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dnamaging}
\alias{df_dnamaging}
\title{data.frame of observations}
\format{A data.frame with 719 rows 3002 columns}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE41037}
}
\usage{
df_dnamaging
}
\description{
A data.frame of 719 observations
}
\keyword{datasets}
