% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{litterature_models}
\alias{litterature_models}
\title{List of litterature models}
\format{A list of litterature models}
\source{
\url{https://doi.org/doi:10.18129/B9.bioc.methylclock}
}
\usage{
litterature_models
}
\description{
A list of 2 litterature models
}
\keyword{datasets}
